backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pymongo>=4.6.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0
pytest>=7.4.0
